Added utilities
